Disallow Content Analysts from editing anything ecosystem-related
