+++ app/controllers/content_analyst/ecosystems_controller.rb
@@ -1,9 +1,3 @@
 class ContentAnalyst::EcosystemsController < ContentAnalyst::BaseController
   include Manager::EcosystemsActions
-
-  protected
-
-  def ecosystems_path
-    content_analyst_ecosystems_path
-  end
 end
