--- app/views/content_analyst/ecosystems/new.html.erb
@@ -1,4 +1,0 @@
-<% @page_header = 'Import a new Ecosystem' %>
-
-<%= render partial: 'manager/ecosystems/import',
-           locals: { ecosystems_path: content_analyst_ecosystems_path } %>
